refactor: split parser and AST into separate files
